Added IsAlive test for blue Api
