+++ BlueApiData/Fixtures/BlueApiTestDataFixture.cs
@@ -54,23 +54,6 @@ namespace BlueApiData.Fixtures
             PrepareTestData().Wait();
         }
 
-        private async Task PrepareTestData()
-        {
-            ApiEndpointNames = new Dictionary<string, string>();
-            ApiEndpointNames["Pledges"] = "/api/pledges";
-
-            _pledgesToDelete = new Dictionary<string, string>();
-
-            TestClientId = _configBuilder.Config["AuthClientId"];
-            TestPledgeCreateClientId = _configBuilder.Config["AuthPledgeCreateClientId"];
-            TestPledgeDeleteClientId = _configBuilder.Config["AuthPledgeDeleteClientId"];
-            TestPledgeUpdateClientId = _configBuilder.Config["AuthPledgeUpdateClientId"];
-
-            TestPledge = await CreateTestPledge(clientId: TestClientId);
-            TestPledgeUpdate = await CreateTestPledge(TestPledgeUpdateClientId, "UpdatePledge");
-            TestPledgeDelete = await CreateTestPledge(TestPledgeDeleteClientId, "DeletePledge");
-        }
-
         private void PrepareDependencyContainer()
         {
             var builder = new ContainerBuilder();
